Gjeninnført teamets AD-gruppe for å kunne kalle import-endepunkt
Samler oppsett av Ktor-testene, for å fjerne duplikat kode - work in progres...
