+++ .nais/prod.yaml
@@ -5,5 +5,6 @@ statistikkTopic: arbeidsgiver.sykefravarsstatistikk-v1
 fiaSuperbrukerGroup: f71a0eed-1e59-4b19-ae13-6c1af824fe8e
 fiaSaksbehandlerGroup: a72547f2-c92b-420c-b7b0-fc227337e9ea
 fiaLesetilgangGroup: 790acd42-dc41-46ce-b7e7-cee414b4b4bc
+teamPiaGroup: 1019baaf-898f-49db-b7b8-e09f0b5b051f
 dbTier: db-custom-2-3840
 dbAutoresizeDisk: true
