+++ src/main/kotlin/no/nav/lydia/NaisEnvironment.kt
@@ -54,7 +54,8 @@ class AzureConfig(
 class FiaRoller(
     val superbrukerGroupId: String = getEnvVar("FIA_SUPERBRUKER_GROUP_ID"),
     val saksbehandlerGroupId: String = getEnvVar("FIA_SAKSBEHANDLER_GROUP_ID"),
-    val lesetilgangGroupId: String = getEnvVar("FIA_LESETILGANG_GROUP_ID")
+    val lesetilgangGroupId: String = getEnvVar("FIA_LESETILGANG_GROUP_ID"),
+    val teamPiaGroupId: String = getEnvVar("TEAM_PIA_GROUP_ID")
 )
 
 class Kafka(
